Moved attributed_text from /selectable_text to /core because its more of a content type than it is a UI concept.
