+++ example/lib/spikes/editor_abstractions/core/attributed_text.dart
