+++ example/lib/spikes/editor_abstractions/default_editor/list_items.dart
@@ -1,10 +1,10 @@
-import 'package:example/spikes/editor_abstractions/selectable_text/attributed_text.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 
+import '../core/attributed_text.dart';
 import '../core/document_composer.dart';
 import '../core/document_editor.dart';
 import '../core/document.dart';
