+++ example/lib/spikes/editor_abstractions/main_inline_styles.dart
@@ -1,4 +1,4 @@
-import 'package:example/spikes/editor_abstractions/selectable_text/attributed_text.dart';
+import 'package:example/spikes/editor_abstractions/core/attributed_text.dart';
 import 'package:example/spikes/editor_abstractions/selectable_text/selectable_text.dart';
 import 'package:flutter/material.dart' hide SelectableText;
